refactored validation, wip
