+++ src/components/auth/controller.ts
@@ -1,16 +1,12 @@
-import { Response } from 'express';
+import { Request, Response } from 'express';
 import { userService } from '@components/user';
-import { ModifiedRequest } from '../../constants/interface';
+import { constants } from '@constants/constatnts';
 import { setDataToRedis } from './services';
-import { constants } from '../../constants/constatnts';
 
-export const signUp = async (req: ModifiedRequest, res: Response): Promise<Response> => {
+export const signUp = async (req: Request, res: Response): Promise<Response> => {
   const { body, sessionID, session } = req;
   const { login, password } = body;
 
-  if (!login || !password) {
-    return res.status(400).send('Bad request');
-  }
   const existingUser = await userService.getUserByLogin(login);
   const hashedPass = await userService.createHashedPassword(password);
 
@@ -27,7 +23,7 @@ export const signUp = async (req: ModifiedRequest, res: Response): Promise<Respo
   }
 };
 
-export const signIn = async (req: ModifiedRequest, res: Response): Promise<Response | void> => {
+export const signIn = async (req: Request, res: Response): Promise<Response | void> => {
   const { body, sessionID, session } = req;
   const { login, password } = body;
