+++ src/components/post/index.ts
@@ -1,2 +1,4 @@
 export * as postInterface from './interfaces';
 export * as postRouter from './router';
+export * as postSchemas from './schemas';
+export { Post } from '../post/models/Post';
