+++ src/components/post/interfaces.ts
@@ -1,8 +1,8 @@
-import { User } from '../user/models/User';
+import { User } from '@components//user';
 
 export interface IPost {
   id?: number;
   title: string;
   text: string;
-  user: User;
+  user?: User;
 }
