+++ src/components/user/index.ts
@@ -1,3 +1,5 @@
 export * as userRouter from './router';
 export * as userService from './services';
 export * as userInterface from './interfaces';
+export * as userSchemas from './schemas';
+export { User } from './models/User';
