+++ src/components/user/router.ts
@@ -1,7 +1,7 @@
 import express from 'express';
+import { validateRequest } from '@utils/requestValidation';
+import { getQuerySchema } from '@utils/schemas';
 import { getAllUsers } from './controller';
-import { validateRequest } from '../../utils/requestValidation';
-import { getQuerySchema } from '../../utils/schemas';
 
 export const user = express.Router();
