+++ src/components/user/schemas.ts
@@ -1,0 +1,8 @@
+import * as zod from 'zod';
+
+export const createUserSchema = {
+  body: zod.object({
+    login: zod.string(),
+    password: zod.string(),
+  }),
+};
